+++ examples/timestamping/backend/src/transactions.rs
@@ -17,8 +17,6 @@
 // Workaround for `failure` see https://github.com/rust-lang-nursery/failure/issues/223 and
 // ECR-1771 for the details.
 #![allow(bare_trait_objects)]
-// Suppress a warning in `transactions!` macro call:
-#![cfg_attr(feature = "cargo-clippy", allow(clippy::redundant_field_names))]
 
 use exonum::{
     blockchain::{ExecutionError, ExecutionResult, Transaction, TransactionContext},