+++ exonum/benches/criterion/block.rs
@@ -114,7 +114,6 @@ mod timestamping {
     use exonum::{
         blockchain::{ExecutionResult, Service, Transaction, TransactionContext},
         crypto::{CryptoHash, Hash, PublicKey, SecretKey},
-        encoding::Error as EncodingError,
         messages::{Message, RawTransaction, Signed},
         storage::Snapshot,
     };
@@ -137,7 +136,7 @@ mod timestamping {
             Vec::new()
         }
 
-        fn tx_from_raw(&self, raw: RawTransaction) -> Result<BoxedTx, EncodingError> {
+        fn tx_from_raw(&self, raw: RawTransaction) -> Result<BoxedTx, failure::Error> {
             use exonum::blockchain::TransactionSet;
             Ok(TimestampingTransactions::tx_from_raw(raw)?.into())
         }
@@ -181,22 +180,12 @@ mod timestamping {
     }
 
     impl Transaction for Tx {
-        fn verify(&self) -> bool {
-            // We don't verify transactions within the benchmark, so in this
-            // and following transaction types the verification code is trivial.
-            unimplemented!("never used in benchmark")
-        }
-
         fn execute(&self, _: TransactionContext) -> ExecutionResult {
             Ok(())
         }
     }
 
     impl Transaction for PanickingTx {
-        fn verify(&self) -> bool {
-            unimplemented!("never used in benchmark")
-        }
-
         fn execute(&self, _: TransactionContext) -> ExecutionResult {
             panic!("panic text");
         }
@@ -225,7 +214,6 @@ mod cryptocurrency {
     use exonum::{
         blockchain::{ExecutionError, ExecutionResult, Service, Transaction, TransactionContext},
         crypto::{Hash, PublicKey, SecretKey},
-        encoding::Error as EncodingError,
         messages::{Message, RawTransaction, Signed},
         storage::{MapIndex, ProofMapIndex, Snapshot},
     };
@@ -253,7 +241,7 @@ mod cryptocurrency {
             Vec::new()
         }
 
-        fn tx_from_raw(&self, raw: RawTransaction) -> Result<BoxedTx, EncodingError> {
+        fn tx_from_raw(&self, raw: RawTransaction) -> Result<BoxedTx, failure::Error> {
             use exonum::blockchain::TransactionSet;
             Ok(CryptocurrencyTransactions::tx_from_raw(raw)?.into())
         }
@@ -337,10 +325,6 @@ mod cryptocurrency {
     }
 
     impl Transaction for Tx {
-        fn verify(&self) -> bool {
-            unimplemented!("never used in benchmark")
-        }
-
         fn execute(&self, mut context: TransactionContext) -> ExecutionResult {
             let from = context.author();
             let mut index = ProofMapIndex::new("provable_balances", context.fork());
@@ -355,10 +339,6 @@ mod cryptocurrency {
     }
 
     impl Transaction for SimpleTx {
-        fn verify(&self) -> bool {
-            unimplemented!("never used in benchmark")
-        }
-
         fn execute(&self, mut context: TransactionContext) -> ExecutionResult {
             let from = context.author();
 
@@ -374,10 +354,6 @@ mod cryptocurrency {
     }
 
     impl Transaction for RollbackTx {
-        fn verify(&self) -> bool {
-            unimplemented!("never used in benchmark")
-        }
-
         fn execute(&self, mut context: TransactionContext) -> ExecutionResult {
             let from = context.author();