+++ exonum/build.rs
@@ -14,7 +14,7 @@ fn create_path_to_protobuf_schema_env() {
     // and dependents in their `build.rs` will have access to `$DEP_EXONUM_PROTOBUF_PROTOS`.
     let path = env::current_dir()
         .expect("Failed to get current dir.")
-        .join("src/encoding/protobuf/proto");
+        .join("src/proto/schema/exonum");
     println!("cargo:protos={}", path.to_str().unwrap());
 }
 
@@ -37,25 +37,22 @@ fn main() {
     create_path_to_protobuf_schema_env();
 
     protobuf_generate(
-        "src/encoding/protobuf/proto/",
-        &["src/encoding/protobuf/proto"],
+        "src/proto/schema/exonum",
+        &["src/proto/schema/exonum"],
         "exonum_proto_mod.rs",
     );
 
     // Exonum external tests.
     protobuf_generate(
         "tests/explorer/blockchain/proto",
-        &[
-            "tests/explorer/blockchain/proto",
-            "src/encoding/protobuf/proto",
-        ],
+        &["tests/explorer/blockchain/proto", "src/proto/schema/exonum"],
         "exonum_tests_proto_mod.rs",
     );
 
     // Exonum benchmarks.
     protobuf_generate(
         "benches/criterion/proto",
-        &["benches/criterion/proto", "src/encoding/protobuf/proto"],
+        &["benches/criterion/proto", "src/proto/schema/exonum"],
         "exonum_benches_proto_mod.rs",
     );
 }